chore : organize file sturcture
