+++ src/components/Post/SinglePost/index.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { baseModalStyle } from 'util/baseStyle';
+
+const Post = () => {};
