+++ src/components/Post/index.ts
@@ -1,0 +1,5 @@
+// export { default as PostList } from './PostList';
+export { default as PreviewPost } from './PreviewPost';
+// export { default as SinglePost } from './SinglePost';
+export { default as WriteButton } from './WriteButton';
+export { default as WritePost } from './WritePost';
